Updates
Many new updates.  A file with details of changes will be added soon.
Version 6.5

+++ Senior Project/SeniorProject/SeniorProject/BoardSpace.cpp
@@ -1,6 +1,7 @@
 #include "BoardSpace.h"
 
-BoardSpace::BoardSpace(){
+BoardSpace::BoardSpace()
+{
 	posX						=	0;
 	posY						=	0;
 	spaceNumber					=	0;
@@ -14,19 +15,22 @@ BoardSpace::BoardSpace(){
 	unitAnimationRect.right		=	0;
 };
 
-void BoardSpace::setPosX(int pX){
+void BoardSpace::setPosX(float pX)
+{
 	posX = pX;
 };
 
-int BoardSpace::getPosX(){
+float BoardSpace::getPosX()
+{
 	return posX;
 };
 
-void BoardSpace::setPosY(int pY){
+void BoardSpace::setPosY(float pY)
+{
 	posY = pY;
 };
 
-int BoardSpace::getPosY(){
+float BoardSpace::getPosY(){
 	return posY;
 };
 
@@ -34,45 +38,56 @@ void BoardSpace::setSpaceNumber(int space){
 	spaceNumber = space;
 };
 
-int BoardSpace::getSpaceNumber(){
+int BoardSpace::getSpaceNumber()
+{
 	return spaceNumber;
 };
 
-void BoardSpace::setOccupiedBy(int type){
+void BoardSpace::setOccupiedBy(int type)
+{
 	occupiedBy = type;
 };
 
-int BoardSpace::getOccupiedBy(){
+int BoardSpace::getOccupiedBy()
+{
 	return occupiedBy;
 };
 
-void BoardSpace::setHighlight(bool b){
+void BoardSpace::setHighlight(bool b)
+{
 	highlighted = b;
 };
 
-bool BoardSpace::isHighlighted(){
+bool BoardSpace::isHighlighted()
+{
 	return highlighted;
 };
 
-void BoardSpace::setUnitCanTakeAction(bool b){
+void BoardSpace::setUnitCanTakeAction(bool b)
+{
 	unitCanTakeAction  = b;
 };
 
-bool BoardSpace::canUnitTakeAction(){
+bool BoardSpace::canUnitTakeAction()
+{
 	return unitCanTakeAction;
 };
 
-void BoardSpace::setAnimationTimer(double time){
+void BoardSpace::setAnimationTimer(double time)
+{
 	animationTimer = time;
 };
 
-void BoardSpace::adjustAnimationTimer(double time){
+void BoardSpace::adjustAnimationTimer(double time)
+{
 	animationTimer += time;
 };
 
-float BoardSpace::getAnimationTimer(){
+double BoardSpace::getAnimationTimer()
+{
 	return animationTimer;
 };
+
 void BoardSpace::setAnimationRect(int t, int l, int r, int b){
 	unitAnimationRect.top		=	t;
 	unitAnimationRect.left		=	l;
